+++ app/src/main/java/com/pozmaxpav/cinemaopinion/domain/repository/firebase/PersonalMovieRepository.kt
@@ -7,14 +7,9 @@ interface PersonalMovieRepository {
     fun removeListener()
     suspend fun addMovieToPersonalList(userId: String, selectedMovie: DomainSelectedMovieModel)
     suspend fun getListPersonalMovies(userId: String, ): List<DomainSelectedMovieModel>
-    suspend fun observeListSelectedMovies(
-        userId: String, onSelectedMoviesUpdated: (List<DomainSelectedMovieModel>) -> Unit
-    )
+    suspend fun observeListSelectedMovies(userId: String, onSelectedMoviesUpdated: (List<DomainSelectedMovieModel>) -> Unit)
     suspend fun deleteMovieFromPersonalList(userId: String, selectedMovieId: Int)
-
     suspend fun addCommentToMovieInPersonalList(userId: String, selectedMovieId: Int, comment: DomainCommentModel)
     suspend fun getCommentsForMovieFromPersonalList(userId: String, selectedMovieId: Int): List<DomainCommentModel>
-    suspend fun observeCommentsForMovieFromPersonalList(
-        userId: String, selectedMovieId: Int, onCommentsUpdated: (List<DomainCommentModel>) -> Unit
-    )
+    suspend fun observeCommentsForMovieFromPersonalList(userId: String, selectedMovieId: Int, onCommentsUpdated: (List<DomainCommentModel>) -> Unit)
 }
