+++ app/src/main/java/com/pozmaxpav/cinemaopinion/domain/repository/firebase/SeriesControlRepository.kt
@@ -6,9 +6,7 @@ interface SeriesControlRepository {
     fun removeListener()
     suspend fun addNewEntry(userId: String, entry: DomainSeriesControlModel)
     suspend fun getListEntries(userId: String): List<DomainSeriesControlModel>
-    suspend fun observeListEntries(
-        userId: String, onEntriesUpdated: (List<DomainSeriesControlModel>) -> Unit
-    )
+    suspend fun observeListEntries(userId: String, onEntriesUpdated: (List<DomainSeriesControlModel>) -> Unit)
     suspend fun deleteEntry(userId: String, entryId: String)
     suspend fun updateEntry(userId: String, entryId: String, selectedEntry: DomainSeriesControlModel)
 }
