+++ app/src/main/java/com/pozmaxpav/cinemaopinion/presentation/navigation/NavGraph.kt
@@ -1,7 +1,6 @@
 package com.pozmaxpav.cinemaopinion.presentation.navigation
 
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
@@ -10,7 +9,6 @@ import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.example.introductoryscreens.ui.onboarding.OnBoardingScreen
-import com.pozmaxpav.cinemaopinion.R
 import com.pozmaxpav.cinemaopinion.presentation.screens.mainScreens.ListOfChangesScreen
 import com.pozmaxpav.cinemaopinion.presentation.screens.mainScreens.ListSelectedGeneralMovies
 import com.pozmaxpav.cinemaopinion.presentation.screens.mainScreens.ListSelectedGeneralSerials
@@ -25,8 +23,8 @@ import com.pozmaxpav.cinemaopinion.presentation.screens.settingsScreens.EditPers
 //import com.pozmaxpav.cinemaopinion.presentation.screens.settingsScreens.EditPersonalInformationScreen
 import com.pozmaxpav.cinemaopinion.presentation.screens.settingsScreens.SettingsScreen
 import com.pozmaxpav.cinemaopinion.presentation.screens.settingsScreens.WebViewScreen
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.ThemeViewModel
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.introduction.OnBoardingViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.system.ThemeViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.system.introduction.OnBoardingViewModel
 
 @Composable
 fun NavGraph(
