+++ app/src/main/java/com/pozmaxpav/cinemaopinion/presentation/screens/mainScreens/ListSelectedGeneralMovies.kt
@@ -55,9 +55,10 @@ import com.pozmaxpav.cinemaopinion.presentation.components.ExpandedCard
 import com.pozmaxpav.cinemaopinion.presentation.components.MyBottomSheet
 import com.pozmaxpav.cinemaopinion.presentation.components.detailscards.DetailsCardSelectedMovie
 import com.pozmaxpav.cinemaopinion.presentation.navigation.Route
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.AuxiliaryUserViewModel
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.FirebaseViewModel
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.MainViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.api.ApiViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.firebase.AuxiliaryUserViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.firebase.FireBaseMovieViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.system.MainViewModel
 import com.pozmaxpav.cinemaopinion.utilits.CustomTextFieldForComments
 import com.pozmaxpav.cinemaopinion.utilits.NODE_LIST_MOVIES
 import com.pozmaxpav.cinemaopinion.utilits.NODE_LIST_WATCHED_MOVIES
@@ -76,9 +77,10 @@ import java.util.Locale
 @Composable
 fun ListSelectedGeneralMovies(
     navController: NavHostController,
-    firebaseViewModel: FirebaseViewModel = hiltViewModel(),
+    firebaseViewModel: FireBaseMovieViewModel = hiltViewModel(),
     auxiliaryUserViewModel: AuxiliaryUserViewModel = hiltViewModel(),
-    mainViewModel: MainViewModel = hiltViewModel()
+    mainViewModel: MainViewModel = hiltViewModel(),
+    apiViewModel: ApiViewModel = hiltViewModel(),
 ) {
     val listMovies by firebaseViewModel.movies.collectAsState()
     val listComments by firebaseViewModel.comments.collectAsState()
@@ -87,7 +89,7 @@ fun ListSelectedGeneralMovies(
     val userId by mainViewModel.userId.collectAsState()
     val userData by auxiliaryUserViewModel.userData.collectAsState()
     val stateMovie by firebaseViewModel.movieDownloadStatus.collectAsState()
-    val info by mainViewModel.informationMovie.collectAsState()
+    val info by apiViewModel.informationMovie.collectAsState()
     val (comment, setComment) = remember { mutableStateOf("") }
     val context = LocalContext.current
     val listState = rememberLazyListState()
@@ -103,7 +105,7 @@ fun ListSelectedGeneralMovies(
 
     LaunchedEffect(selectedMovie) {
         selectedMovie?.let { movie ->
-            mainViewModel.getInformationMovie(movie.id)
+            apiViewModel.getInformationMovie(movie.id)
         }
     }
 
@@ -402,7 +404,7 @@ fun ListSelectedGeneralMovies(
 fun ShowCommentGeneralList(
     listComments: List<DomainCommentModel>,
     id: Double,
-    firebaseViewModel: FirebaseViewModel = hiltViewModel(),
+    firebaseViewModel: FireBaseMovieViewModel = hiltViewModel(),
 ) {
     val stateComments by firebaseViewModel.commentsDownloadStatus.collectAsState()
