+++ app/src/main/java/com/pozmaxpav/cinemaopinion/presentation/screens/settingsScreens/WebViewScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.WebViewViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.system.WebViewViewModel
 
 @SuppressLint("SetJavaScriptEnabled")
 @Composable
