+++ app/src/main/java/com/pozmaxpav/cinemaopinion/utilits/CheckVersionApp.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.pozmaxpav.cinemaopinion.presentation.viewModel.MainViewModel
+import com.pozmaxpav.cinemaopinion.presentation.viewModel.system.MainViewModel
 
 @Composable
 fun CheckAndUpdateAppVersion(
